requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
python-dotenv==1.0.1
openai==1.32.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
pytest==8.2.0
pytest-asyncio==0.23.7
